refactor: 리뷰,마이페이지,메인페이지
