+++ client/reviewRegister/accessible.tsx
@@ -1,24 +1,23 @@
 import { useMemo } from 'react';
 import useNoDuplicateSelect from 'hooks/useNoDuplicateSelect';
 import useReviewRegisterReudx from 'hooks/useReviewRegisterRedux';
-import { dispatchLandlord } from 'store/modules/reviewRegister';
-import NoDuplicateSelectBtn from '../NoDuplicateSelectBtn';
+import { dispatchAccessible } from 'store/modules/reviewRegister';
+import NoDuplicateSelectBtn from 'hooks/reviewRegisterNoDuplicateSelectBtn';
 
-export default function Landlord() {
-  const useTypeArr = useMemo(() => ['만족', '보통', '불만족'], []);
+function Accessible() {
+  const typeArr = useMemo(() => ['만족', '보통', '불만족'], []);
   const [state, dispatch] = useReviewRegisterReudx();
-
   const [selectArr, checkHandler] = useNoDuplicateSelect(
-    useTypeArr,
-    dispatchLandlord,
-    state.landlord,
+    typeArr,
+    dispatchAccessible,
+    state.accessible,
   );
 
   return (
     <div className="mt-[8px] flex text-[15px] Pretendard-Regular items-center justify-between">
-      <div className="w-1/4">집주인 친절</div>
+      <div className="w-1/4">위치</div>
       <div className="grid grid-cols-3 w-3/4">
-        {useTypeArr.map((item, index) => (
+        {typeArr.map((item, index) => (
           <NoDuplicateSelectBtn
             key={index}
             value={item}
@@ -31,3 +30,5 @@ export default function Landlord() {
     </div>
   );
 }
+
+export default Accessible;
