+++ client/reviewRegister/clean.tsx
@@ -1,24 +1,23 @@
 import { useMemo } from 'react';
 import useNoDuplicateSelect from 'hooks/useNoDuplicateSelect';
 import useReviewRegisterReudx from 'hooks/useReviewRegisterRedux';
-import { dispatchFacility } from 'store/modules/reviewRegister';
-import NoDuplicateSelectBtn from '../NoDuplicateSelectBtn';
+import { dispatchClean } from 'store/modules/reviewRegister';
+import NoDuplicateSelectBtn from 'hooks/reviewRegisterNoDuplicateSelectBtn';
 
-export default function Facility() {
-  const useTypeArr = useMemo(() => ['만족', '보통', '불만족'], []);
+function Clean() {
+  const typeArr = useMemo(() => ['만족', '보통', '불만족'], []);
   const [state, dispatch] = useReviewRegisterReudx();
-
   const [selectArr, checkHandler] = useNoDuplicateSelect(
-    useTypeArr,
-    dispatchFacility,
-    state.facility,
+    typeArr,
+    dispatchClean,
+    state.clean,
   );
 
   return (
-    <div className="mt-[8px] pb-[30px] flex text-[15px] Pretendard-Regular items-center justify-between border-b border-border_color">
-      <div className="w-1/4">시설 만족도</div>
+    <div className="mt-[45px] flex text-[15px] Pretendard-Regular items-center justify-between">
+      <div className="w-1/4">청결도</div>
       <div className="grid grid-cols-3 w-3/4">
-        {useTypeArr.map((item, index) => (
+        {typeArr.map((item, index) => (
           <NoDuplicateSelectBtn
             key={index}
             value={item}
@@ -31,3 +30,5 @@ export default function Facility() {
     </div>
   );
 }
+
+export default Clean;
