+++ client/reviewRegister/facility.tsx
@@ -1,24 +1,23 @@
 import { useMemo } from 'react';
 import useNoDuplicateSelect from 'hooks/useNoDuplicateSelect';
 import useReviewRegisterReudx from 'hooks/useReviewRegisterRedux';
-import { dispatchAccessible } from 'store/modules/reviewRegister';
-import NoDuplicateSelectBtn from '../NoDuplicateSelectBtn';
+import { dispatchFacility } from 'store/modules/reviewRegister';
+import NoDuplicateSelectBtn from 'hooks/reviewRegisterNoDuplicateSelectBtn';
 
-export default function Accessible() {
-  const typeArr = useMemo(() => ['만족', '보통', '불만족'], []);
+function Facility() {
+  const useTypeArr = useMemo(() => ['만족', '보통', '불만족'], []);
   const [state, dispatch] = useReviewRegisterReudx();
-
   const [selectArr, checkHandler] = useNoDuplicateSelect(
-    typeArr,
-    dispatchAccessible,
-    state.accessible,
+    useTypeArr,
+    dispatchFacility,
+    state.facility,
   );
 
   return (
-    <div className="mt-[8px] flex text-[15px] Pretendard-Regular items-center justify-between">
-      <div className="w-1/4">위치</div>
+    <div className="mt-[8px] pb-[30px] flex text-[15px] Pretendard-Regular items-center justify-between border-b border-border_color">
+      <div className="w-1/4">시설 만족도</div>
       <div className="grid grid-cols-3 w-3/4">
-        {typeArr.map((item, index) => (
+        {useTypeArr.map((item, index) => (
           <NoDuplicateSelectBtn
             key={index}
             value={item}
@@ -31,3 +30,5 @@ export default function Accessible() {
     </div>
   );
 }
+
+export default Facility;
