+++ client/reviewRegister/period.tsx
@@ -1,14 +1,19 @@
+import { ChangeEvent } from 'react';
 import useReviewRegisterReudx from 'hooks/useReviewRegisterRedux';
-import {
-  dispatchPeriodFrom,
-  dispatchPeriodTo,
-} from '../../../store/modules/reviewRegister';
+import {dispatchPeriodFrom, dispatchPeriodTo} from '../../store/modules/reviewRegister';
 
-export default function MoveDate() {
+function MoveDate() {
   const [state, dispatch] = useReviewRegisterReudx();
 
   const periodFrom = state.periodFrom;
   const periodTo = state.periodTo;
+
+  const getPeriodFrom = (e: ChangeEvent<HTMLInputElement>) => {
+    dispatch(dispatchPeriodFrom(e.target.value))
+  };
+  const getPeriodTo = (e: ChangeEvent<HTMLInputElement>) => {
+    dispatch(dispatchPeriodTo(e.target.value))
+  }
 
   return (
     <>
@@ -16,21 +21,19 @@ export default function MoveDate() {
         <input
           type="date"
           className="text-center Pretendard-Regular bg-white  w-full rounded-standard_rounded h-full text-[15px] px-4 placeholder-black outline-0 "
-          onChange={(e) => {
-            dispatch(dispatchPeriodFrom(e.target.value));
-          }}
+          onChange={(e) => getPeriodFrom(e)}
           value={periodFrom}
         />
         ~
         <input
           type="date"
           className="text-center Pretendard-Regular bg-white  w-full rounded-standard_rounded h-full text-[15px] px-4 placeholder-black outline-0 "
-          onChange={(e) => {
-            dispatch(dispatchPeriodTo(e.target.value));
-          }}
+          onChange={(e) => getPeriodTo(e)}
           value={periodTo}
         />
       </div>
     </>
   );
 }
+
+export default MoveDate;
