+++ client/reviewRegister/subHeader.tsx
@@ -1,16 +1,18 @@
 import useReviewRegisterRedux from 'hooks/useReviewRegisterRedux';
 import { componentHandle } from 'store/modules/reviewRegister';
 
-export default function SubHeader() {
+const subHeaderBtnArr = ['주소 등록', '후기 작성'];
+
+function SubHeader() {
   const [state, dispatch] = useReviewRegisterRedux();
 
   const onClickHandler = (num: number) => {
     dispatch(componentHandle(num));
   };
-  const btnArr = ['주소 등록', '후기 작성'];
+  
   return (
     <div className="sm:w-[375px] sm:m-auto text-font_gray flex w-full justify-around px-[18px] bg-component_white items-center border-b border-border_color">
-      {btnArr.map((title, index) => (
+      {subHeaderBtnArr.map((title, index) => (
         <button
           className={`py-4 px-20 text-[14px] Pretendard-Regular  h-max ${
             state.COMPONENT_HANDLER === index &&
@@ -25,3 +27,5 @@ export default function SubHeader() {
     </div>
   );
 }
+
+export default SubHeader;
