+++ package.json
@@ -21,7 +21,7 @@
     "next-redux-saga": "^4.1.2",
     "next-redux-wrapper": "^7.0.5",
     "postcss": "^7.0.39",
-    "rc-slider": "^10.0.0-alpha.6",
+    "rc-slider": "^10.0.0-alpha.6", 
     "react": "^17.0.2",
     "react-daum-postcode": "^3.1.1",
     "react-dom": "^17.0.2",
